requirements.txt
pytest
requests
flask
flask-sqlalchemy
flask-admin
python-dateutil